+++ README.md
@@ -3,6 +3,7 @@
 ## Installation
 
 ### Install Core Depedencies
+
 ```bash
 pip install -r requirements.txt
 pre-commit install
@@ -10,9 +11,11 @@ pip install -e ./src/core
 ```
 
 ### Install modules you want to work on
+
 ```bash
 pip install -e ./src/<module_name>
 ```
+
 ```bash
 pip install -e ./src/data
 pip install -e ./src/model
@@ -20,10 +23,9 @@ pip install -e ./src/evaluation
 ```
 
 ## Testing
+
 To Run Tests on Development Environment
-```bash
-pip install -e ./tests/<module_name>
-```
+
 ```bash
 pytest tests/data
 pytest tests/model
