--- docker/Dockerfile
@@ -1,2 +1,0 @@
-FROM pytorch/pytorch:1.12.1-cuda11.6-cudnn8-runtime
-RUN apt-get update
