--- src/model/openthaigpt_pretraining_model/llama/generation.py
@@ -1,1 +1,0 @@
-from llama.generation import LLaMA  # type: ignore # noqa: F401
