--- src/model/openthaigpt_pretraining_model/llama/tokenizer.py
@@ -1,1 +1,0 @@
-from llama.tokenizer import Tokenizer  # type: ignore  # noqa: F401
