+++ src/model/openthaigpt_pretraining_model/models/llama/generation.py
@@ -1,0 +1,1 @@
+from llama.generation import LLaMA  # type: ignore # noqa: F401
