+++ src/model/openthaigpt_pretraining_model/models/llama/tokenizer.py
@@ -1,0 +1,1 @@
+from llama.tokenizer import Tokenizer  # type: ignore  # noqa: F401
